Extract parsers to a new file
