+++ client/src/containers/section/page/networks/network-detail-panel.tsx
@@ -3,7 +3,6 @@
 import { ArrowLeft, ExternalLink, Pencil } from 'lucide-react';
 
 import { cn } from '@/lib/classnames';
-import { format } from '@/lib/utils/formats';
 
 import { Organization, Project } from '@/types/generated/strapi.schemas';
 
@@ -12,9 +11,12 @@ import { NetworkResponse } from '@/hooks/networks';
 import type { OpenDetails } from '@/containers/section/page/pages/network';
 
 import { SlidingButton, Button } from '@/components/ui/button';
-type ProjectWithType = { type: 'project'; attributes: Project; id: number };
 
-type OrganizationWithType = { type: 'organization'; attributes: Organization; id: number };
+import { getProjectFields, getOrganizationFields } from './network-detail-parsers';
+
+export type ProjectWithType = { type: 'project'; attributes: Project; id: number };
+
+export type OrganizationWithType = { type: 'organization'; attributes: Organization; id: number };
 
 const isProject = (
   dataWithType: Data & { type: Type },
@@ -30,115 +32,6 @@ type Field = {
   label: string;
   value: string | (string | undefined)[] | undefined;
   url?: string | string[];
-};
-
-const getOrganizationFields = (dataWithType: OrganizationWithType) => {
-  const {
-    country,
-    main_organization_theme: thematic,
-    secondary_organization_theme: secondaryThematic,
-    organization_type: organizationType,
-  } = dataWithType.attributes || {};
-  const fields = [];
-  if (typeof country !== 'undefined') {
-    fields.push({ label: 'Country', value: country?.data?.attributes?.name });
-  }
-  if (typeof thematic !== 'undefined') {
-    const thematics = `${thematic?.data?.attributes?.name}${secondaryThematic?.data?.attributes?.name
-        ? `, ${secondaryThematic?.data?.attributes?.name}`
-        : ''
-      } `;
-    fields.push({ label: 'Thematic', value: thematics });
-  }
-  if (typeof organizationType !== 'undefined') {
-    fields.push({
-      label: 'Type of organisation',
-      value: organizationType?.data?.attributes?.name,
-    });
-  }
-  return fields;
-};
-
-const hasData = (field: Project[keyof Project]) => {
-  if (!field || field === '') return false;
-  if (typeof field === 'string') return true;
-  return typeof field !== 'undefined' && field?.data;
-};
-
-const getProjectFields = (dataWithType: ProjectWithType) => {
-  const {
-    start_date: startDate,
-    end_date: endDate,
-    country_of_coordination: countryOfCoordination,
-    project_coordinator_name: projectCoordinatorName,
-    project_coordinator_email: projectCoordinatorEmail,
-    second_project_coordinator_name: secondProjectCoordinatorName,
-    second_project_coordinator_email: secondProjectCoordinatorEmail,
-    project_type: projectType,
-    country_of_interventions: countryOfInterventions,
-    main_area_of_intervention: mainAreaOfIntervention,
-    sustainable_development_goal: sustainableDevelopmentGoal,
-  } = dataWithType.attributes || {};
-
-  const fields = [];
-  if (typeof startDate !== 'undefined') {
-    const formatDate = (date: string) =>
-      format({
-        id: 'formatDate',
-        value: date,
-      });
-
-    fields.push({
-      label: 'Dates',
-      value: `${formatDate(startDate)}${endDate ? ` - ${formatDate(endDate)}` : ''}`,
-    });
-  }
-  if (hasData(countryOfCoordination)) {
-    fields.push({
-      label: 'Country of coordination',
-      value: countryOfCoordination?.data?.attributes?.name,
-    });
-  }
-  if (hasData(projectCoordinatorName)) {
-    if (hasData(secondProjectCoordinatorName)) {
-      fields.push({
-        label: 'Project coordinators',
-        value: [projectCoordinatorName, secondProjectCoordinatorName],
-        url: [`mailto:${projectCoordinatorEmail}`, `mailto:${secondProjectCoordinatorEmail}`],
-      });
-    } else {
-      fields.push({
-        label: 'Project coordinator',
-        value: projectCoordinatorName,
-        url: `mailto:${projectCoordinatorEmail}`,
-      });
-    }
-  }
-  if (hasData(projectType)) {
-    fields.push({
-      label: 'Project type',
-      value: projectType?.data?.attributes?.name,
-    });
-  }
-  if (hasData(countryOfInterventions) && countryOfInterventions?.data?.length) {
-    fields.push({
-      label: 'Country of interventions',
-      value: countryOfInterventions?.data?.map((c) => c.attributes?.name).join(', '),
-    });
-  }
-  if (hasData(mainAreaOfIntervention)) {
-    fields.push({
-      label: 'Main area of intervention',
-      value: mainAreaOfIntervention?.data?.attributes?.name,
-    });
-  }
-  if (hasData(sustainableDevelopmentGoal)) {
-    fields.push({
-      label: 'Sustainable Development Goal',
-      value: sustainableDevelopmentGoal?.data?.attributes?.name,
-    });
-  }
-  return fields;
 };
 
 const Field = ({ label, value, url, type }: Field & { type: Type }) => {
